+++ src/main.cpp
@@ -13,8 +13,10 @@
 #include <QApplication>
 #include <QCommandLineParser>
 #include <QDebug>
+#include <QDir>
 #include <QStyleFactory>
 
+#include "./core/configmanager.h"
 #include "./ui/mainwindow.h"
 
 int main(int argc, char *argv[])
@@ -28,10 +30,28 @@ int main(int argc, char *argv[])
 
   QCommandLineParser parser;
   parser.addHelpOption();
+  parser.addVersionOption();
   parser.addPositionalArgument("source", "The source file to open.");
+  QCommandLineOption configFileOption("c", "specify configuration file.",
+                                      "config.ini");
+  parser.addOption(configFileOption);
   parser.process(App);
 
-  MainWindow notepanda;
+  QString configFile = parser.value(configFileOption);
+  if (configFile.isEmpty()) {
+#ifdef Q_OS_WIN
+    configFile = App.applicationDirPath() + "/config.ini";
+#else
+    QDir configDir = QDir::homePath() + "/.config/notepanda";
+    configFile = configDir.absolutePath() + "/config.ini";
+    if (!configDir.exists()) {
+      configDir.mkpath(configDir.absolutePath());
+    }
+#endif
+  }
+  ConfigManager *configManager;
+  configManager = new ConfigManager(configFile);
+  MainWindow notepanda(configManager);
   notepanda.show();
   if (parser.positionalArguments().size() == 1)
     notepanda.plainTextEdit->openFile(parser.positionalArguments().at(0));