+++ src/ui/preferenceswindow.cpp
@@ -19,31 +19,19 @@
 #include "../core/texteditor.h"
 #include "ui_preferenceswindow.h"
 
-PreferencesWindow::PreferencesWindow(QWidget *parent, TextEditor *pTE)
-    : QDialog(parent), ui(new Ui::PreferencesWindow), plainTextEdit(pTE)
+PreferencesWindow::PreferencesWindow(ConfigManager *cfManager, QWidget *parent)
+    : QDialog(parent), ui(new Ui::PreferencesWindow), configManager(cfManager)
 {
   ui->setupUi(this);
-  configManager = new ConfigManager;
   setWindowTitle(tr("Preferences - Notepanda"));
 
+  resetAllValues();
+}
+
+void PreferencesWindow::resetAllValues()
+{
   ui->themeCombo->addItems(QStyleFactory::keys());
   ui->themeCombo->setCurrentText(configManager->getStyle());
   ui->fontComboBox->setCurrentFont(QFont(configManager->getEditorFontFamily()));
   ui->spinBox->setValue(configManager->getEditorFontSize());
-
-  connect(ui->themeCombo, &QComboBox::currentTextChanged, this,
-          [&](const QString &curTheme) {
-            QApplication::setStyle(QStyleFactory::create(curTheme));
-            configManager->setStyle(curTheme);
-          });
-  connect(ui->fontComboBox, &QFontComboBox::currentFontChanged, this,
-          [this](const QFont font) { plainTextEdit->setEditorFont(font); });
-  connect(ui->spinBox, QOverload<int>::of(&QSpinBox::valueChanged),
-          [=](const int &value) {
-            plainTextEdit->setEditorFontSize(value);
-            qDebug() << "in PreferencesWindow"
-                     << configManager->getEditorFontSize();
-          });
-  connect(ui->buttonBox, &QDialogButtonBox::accepted, this,
-          [&]() { configManager->save(); });
 }